feat: ワークフローと機能ベースのディレクトリ構成に移行 - financial_months
