--- backend/src/features/financial_months/infrastructure/entity.ts
@@ -1,12 +1,0 @@
-import { z } from 'zod'
-
-export const FinancialMonthRecord = z.object({
-  id: z.string(),
-  user_id: z.string(),
-  financial_year: z.number(),
-  month: z.number(),
-  started_at: z.number(),
-  ended_at: z.number(),
-})
-
-export type FinancialMonthRecord = z.infer<typeof FinancialMonthRecord>
